refactor: AvailableVersionText
